Everything until the calculation process have been implemented! now need to use the docker as well as complete the left part of the calculation payment method in serviceImpl!

+++ src/main/java/by/kuvonchbekn/outlaysbot/service/impl/ProductServiceImpl.java
@@ -3,6 +3,7 @@ package by.kuvonchbekn.outlaysbot.service.impl;
 import by.kuvonchbekn.outlaysbot.assemblers.ProductAssembler;
 import by.kuvonchbekn.outlaysbot.dto.ProductDto;
 import by.kuvonchbekn.outlaysbot.entity.Product;
+import by.kuvonchbekn.outlaysbot.exception.specificExceptions.ProductAlreadyExistsException;
 import by.kuvonchbekn.outlaysbot.exception.specificExceptions.ProductNotFoundException;
 import by.kuvonchbekn.outlaysbot.repo.ProductRepo;
 import by.kuvonchbekn.outlaysbot.service.ProductService;
@@ -23,13 +24,22 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepo productRepo;
     private final MessageSource messageSource;
     @Override
-    public void deleteProduct(String productId) {
-        productRepo.deleteById(productId);
+    public boolean deleteProduct(String productId) {
+        Optional<Product> byId = productRepo.findById(productId);
+        try {
+            if (byId.isEmpty()) throw new ProductNotFoundException(messageSource.getMessage("api.error.product.not.found", null, Locale.ENGLISH));
+            productRepo.deleteById(productId);
+            return true;
+        }catch (ProductNotFoundException e){
+            return false;
+        }
     }
 
     @Override
-    public void saveProduct(Product product) {
-        productRepo.save(product);
+    public Product saveProduct(Product product) {
+        Optional<Product> byName = productRepo.findByName(product.getName());
+        if (byName.isPresent()) throw new ProductAlreadyExistsException(messageSource.getMessage("api.error.product.already.exists", null, Locale.ENGLISH));
+        return productRepo.save(product);
     }
 
     @Override
@@ -38,7 +48,6 @@ public class ProductServiceImpl implements ProductService {
         if (byId.isEmpty()) throw new ProductNotFoundException(messageSource.getMessage("api.error.product.not.found", null, Locale.ENGLISH));
         return byId.get();
     }
-    //need to implement assembler in productController
 
     @Override
     public List<Product> getAllProductsList() {
@@ -46,7 +55,14 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void updateProduct(String productId, ProductDto productDto) {
+    public Product updateProduct(String productId, ProductDto productDto) {
+        Optional<Product> byId = productRepo.findById(productId);
+        if (byId.isEmpty()) throw new ProductNotFoundException(messageSource.getMessage("api.error.product.not.found", null, Locale.ENGLISH));
+        Product productSaved = byId.get();
 
+        productSaved.setName(productDto.getName());
+        productSaved.setPrice(productDto.getPrice());
+
+        return productRepo.save(productSaved);
     }
 }
